Fix other warnings

+++ src/EntityManager.cpp
@@ -8,5 +8,5 @@ void EntityManager::clean_up()
 {
 	for (size_t i{ 0 }; i < entities.size(); i++)
 		if (entities[i]->toBeKilled())
-			entities.erase(entities.begin() + i);
+			entities.erase(entities.begin() + static_cast<long>(i));
 }
